requirements.txt
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
datasets>=2.14.0
faiss-gpu>=1.7.2
pyarrow>=12.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
umap-learn>=0.5.3
tensorboard>=2.13.0
